Completed Sheet Rendering + Adding Clefs
I have completed the basic Sheet class, now letting the user construct a Sheet class that is capable of building and rendering an entire sheet music with the only html necessary to write is importing the necessary html scripts. I have shown this in the updated dic_test.html file, and also made fixed the code to be portable to all html compilers. 
Fixed previous issue of being unable to add the first measure and noteset (was only able to build upon previous before), and now fixed the noteset and measures array to stack into a psuedo-2d array in javascript. The newMeasure() function is now able to add clefs in any created measure.