Dynamics, Multi-clef, Bars, + bug fixes
Continued to add dynamic features including the decrescendo, and text options (for anything like writing fortissimo, or other markings). Fixed bugs to allow for multi-clef, also fixed the rendering of crescendos to include on a stave that does not render, and subtract the y-positions that this invisible stave takes up for other staves. Attachment bars across notes less than a quarter note has also been added as a feature as part of the render function.